Game Queue
- Added in new UI to be able to join and leave a queue
- New timer for time in queue
- New network tags for queue and validation areas
- Fire an event on validationComplete to let main menu have ref to be able to turn on play/join queue

+++ RPG-Unity2DChallenge/Assets/Code/Utility/MethodExtensions.cs
@@ -6,7 +6,7 @@ namespace Project.Utility {
     public static class MethodExtensions {
         
         //Set alpha of a colour property
-		public static Color SetAlpha(this Color C, int Alpha) {
+		public static Color SetAlpha(this Color C, float Alpha) {
             Alpha = Mathf.Clamp(Alpha, 0, 1);
             C.a = Alpha;
             return C;
